+++ src/compas_timber/ghpython/components/CT_SurfaceModelOptions/code.py
@@ -1,3 +1,4 @@
+import Grasshopper
 from ghpythonlib.componentbase import executingcomponent as component
 
 
@@ -24,4 +25,4 @@ class SurfaceModelOptions(component):
             "joint_overrides": joint_overrides,
         }
 
-        return (dict,)
+        return Grasshopper.Kernel.Types.GH_ObjectWrapper(dict)